Version 11.6.5-pre

+++ Src/Include/xxh3.h
@@ -351,7 +351,7 @@ static XXH64_hash_t XXH3_avalanche(uint64 h64)
 XXH_FORCE_INLINE XXH64_hash_t XXH3_len_1to3_64b(const void * data, size_t len, const void * keyPtr, XXH64_hash_t seed)
 {
 	assert(data);
-	assert(checkirange(len, 1, 3));
+	assert(checkirange(len, static_cast<size_t>(1), static_cast<size_t>(3)));
 	assert(keyPtr);
 	{   
 		BYTE const c1 = ((const BYTE *)data)[0];
@@ -368,7 +368,7 @@ XXH_FORCE_INLINE XXH64_hash_t XXH3_len_4to8_64b(const void * data, size_t len, c
 {
 	assert(data);
 	assert(keyPtr);
-	assert(checkirange(len, 4, 8));
+	assert(checkirange(len, static_cast<size_t>(4), static_cast<size_t>(8)));
 	{   
 		const uint32 in1 = XXH_readLE32(data);
 	    const uint32 in2 = XXH_readLE32((const BYTE *)data + len - 4);
@@ -383,7 +383,7 @@ XXH_FORCE_INLINE XXH64_hash_t XXH3_len_9to16_64b(const void * data, size_t len, 
 {
 	assert(data);
 	assert(keyPtr);
-	assert(checkirange(len, 9, 16));
+	assert(checkirange(len, static_cast<size_t>(9), static_cast<size_t>(16)));
 	{   
 		const uint64* const key64 = (const uint64*)keyPtr;
 	    uint64 const ll1 = XXH_readLE64(data) ^ (XXH_readLE64(key64) + seed);
@@ -1131,7 +1131,7 @@ XXH_PUBLIC_API XXH64_hash_t XXH3_64bits_digest(const XXH3_state_t* state)
 XXH_FORCE_INLINE XXH128_hash_t XXH3_len_1to3_128b(const void * data, size_t len, const void * keyPtr, XXH64_hash_t seed)
 {
 	assert(data);
-	assert(checkirange(len, 1, 3));
+	assert(checkirange(len, static_cast<size_t>(1), static_cast<size_t>(3)));
 	assert(keyPtr);
 	{   
 		const uint32 * const key32 = (const uint32 *)keyPtr;
@@ -1152,7 +1152,7 @@ XXH_FORCE_INLINE XXH128_hash_t XXH3_len_4to8_128b(const void * data, size_t len,
 {
 	assert(data);
 	assert(keyPtr);
-	assert(checkirange(len, 4, 8));
+	assert(checkirange(len, static_cast<size_t>(4), static_cast<size_t>(8)));
 	{   
 		const uint32 in1 = XXH_readLE32(data);
 	    const uint32 in2 = XXH_readLE32((const BYTE *)data + len - 4);
@@ -1175,7 +1175,7 @@ XXH_FORCE_INLINE XXH128_hash_t XXH3_len_9to16_128b(const void * data, size_t len
 {
 	assert(data);
 	assert(keyPtr);
-	assert(checkirange(len, 9, 16));
+	assert(checkirange(len, static_cast<size_t>(9), static_cast<size_t>(16)));
 	{   const uint64* const key64 = (const uint64*)keyPtr;
 	    uint64 const ll1 = XXH_readLE64(data) ^ (XXH_readLE64(key64) + seed);
 	    uint64 const ll2 = XXH_readLE64((const BYTE *)data + len - 8) ^ (XXH_readLE64(key64+1) - seed);
